requirements.txt
flask
flask_cors
flask_sqlalchemy
sqlalchemy < 1.4.0
psycopg2-binary
bcrypt
pyjwt